backend/requirements.txt
fastapi
uvicorn
sqlalchemy
google-auth-oauthlib
google-api-python-client
anthropic
python-dotenv
httpx
